# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Profile API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/user-profiles": {
            "get": {
                "tags": [
                    "User Profiles"
                ],
                "summary": "Get all user profiles",
                "operationId": "8e24f07c7e22849fef609b792da9e058",
                "responses": {
                    "200": {
                        "description": "A list of user profiles"
                    }
                }
            },
            "post": {
                "tags": [
                    "User Profiles"
                ],
                "summary": "Create a new user profile",
                "operationId": "6a36e7ba5213abe5e7517bd5ef1d2b2b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "email",
                                    "phone_number",
                                    "firstName",
                                    "lastName",
                                    "gender"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "gdpr_sms_active": {
                                        "type": "boolean"
                                    },
                                    "gdpr_email_active": {
                                        "type": "boolean"
                                    },
                                    "referred_by": {
                                        "type": "string"
                                    },
                                    "preferred_location": {
                                        "type": "integer"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User profile created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "type": "integer"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "phone_number": {
                                            "type": "string"
                                        },
                                        "firstName": {
                                            "type": "string"
                                        },
                                        "lastName": {
                                            "type": "string"
                                        },
                                        "gender": {
                                            "type": "string"
                                        },
                                        "gdpr_sms_active": {
                                            "type": "boolean"
                                        },
                                        "gdpr_email_active": {
                                            "type": "boolean"
                                        },
                                        "referred_by": {
                                            "type": "string"
                                        },
                                        "preferred_location": {
                                            "type": "integer"
                                        },
                                        "avatar": {
                                            "type": "string"
                                        },
                                        "address": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/user-profiles/{id}": {
            "get": {
                "tags": [
                    "User Profiles"
                ],
                "summary": "Get a single user profile",
                "operationId": "658d6f878dd67b7351bdf96322c410be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile details"
                    },
                    "404": {
                        "description": "User profile not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "User Profiles"
                ],
                "summary": "Update a user profile",
                "operationId": "5ca0c7d4a627a2b06c1d5f0c8fed28ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone_number": {
                                        "type": "string"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "gdpr_sms_active": {
                                        "type": "boolean"
                                    },
                                    "gdpr_email_active": {
                                        "type": "boolean"
                                    },
                                    "referred_by": {
                                        "type": "string"
                                    },
                                    "preferred_location": {
                                        "type": "integer"
                                    },
                                    "avatar": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User profile updated successfully"
                    },
                    "404": {
                        "description": "User profile not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "User Profiles"
                ],
                "summary": "Delete a user profile",
                "operationId": "43176af0153a5990af1e87fa66bd624d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile deleted successfully"
                    },
                    "404": {
                        "description": "User profile not found"
                    }
                }
            }
        },
...
    },
...
}